fix bugs

+++ HW4/OleksandrPanchuk/task.py
@@ -1,4 +1,4 @@
-Celsius = int(input("Enter the temperature in Celsius: "))
+Celsius = float(input("Enter the temperature in Celsius: "))
 Fahrenheit = float((Celsius *9/5) + 32)
 if Celsius <= -273.15:
     print("Error: Temperature below absolute zero (-273.15°C)")
